Create relayer stats endpoint
