+++ src/stats/get-relayer-stats.js
@@ -1,85 +1,26 @@
 const _ = require('lodash');
+const BigNumber = require('bignumber.js');
 
-const { ZRX_TOKEN_ADDRESS } = require('../constants');
-const Fill = require('../model/fill');
-const formatTokenAmount = require('../tokens/format-token-amount');
-const getRelayers = require('../relayers/get-relayers');
-const getTokens = require('../tokens/get-tokens');
+const getRelayersStats = require('./get-relayers-stats');
 
 const getRelayerStats = async (dateFrom, dateTo) => {
-  const metrics = await Fill.aggregate([
-    {
-      $match: {
-        date: { $gte: new Date(dateFrom), $lte: new Date(dateTo) },
-        relayerId: { $ne: null },
-      },
-    },
-    {
-      $group: {
-        _id: {
-          relayerId: `$relayerId`,
-        },
-        localisedMakerFees: { $sum: `$conversions.USD.makerFee` },
-        localisedTakerFees: { $sum: `$conversions.USD.takerFee` },
-        makerFee: { $sum: '$makerFee' },
-        takerFee: { $sum: '$takerFee' },
-        trades: { $sum: 1 },
-        volume: { $sum: `$conversions.USD.amount` },
-      },
-    },
-  ]);
-
-  const relayers = await getRelayers();
-  const tokens = await getTokens();
-  const zrxToken = tokens[ZRX_TOKEN_ADDRESS];
-
-  const stats = _.map(relayers, relayer => {
-    const metric = _.find(metrics, { _id: { relayerId: relayer.lookupId } });
-
-    if (_.isUndefined(metric)) {
-      return {
-        fees: { USD: 0, ZRX: 0 },
-        relayer: relayer.id,
-        trades: 0,
-        volume: 0,
-      };
-    }
-
-    const {
-      localisedMakerFees,
-      localisedTakerFees,
-      makerFee,
-      takerFee,
-      trades,
-      volume,
-    } = metric;
-
-    const stat = {
+  const relayerStats = await getRelayersStats(dateFrom, dateTo);
+  const stats = _.reduce(
+    relayerStats,
+    (acc, stat) => ({
       fees: {
-        USD: localisedMakerFees + localisedTakerFees,
-        ZRX: formatTokenAmount(makerFee + takerFee, zrxToken),
+        USD: acc.fees.USD + stat.fees.USD,
+        ZRX: acc.fees.ZRX.plus(stat.fees.ZRX.toString()),
       },
-      relayer: relayer.id,
-      trades,
-      volume,
-    };
-
-    if (
-      _.includes(
-        ['paradex', 'ddex', 'theOcean', 'starBit', 'tokenmom'],
-        relayer.id,
-      )
-    ) {
-      return {
-        fees: stat.fees,
-        relayer: relayer.id,
-        trades: Math.ceil(stat.trades / 2),
-        volume: stat.volume / 2,
-      };
-    }
-
-    return stat;
-  });
+      trades: acc.trades + stat.trades,
+      volume: acc.volume + stat.volume,
+    }),
+    {
+      fees: { USD: 0, ZRX: new BigNumber(0) },
+      trades: 0,
+      volume: 0,
+    },
+  );
 
   return stats;
 };
